backend/services/news_service.py
import requests
from bs4 import BeautifulSoup
import logging

logger = logging.getLogger(__name__)


class NewsService:
    @staticmethod
    def fetch_transport_news():
        """
        Web Scraping: Budapest Airport hírek és közlemények gyűjtése.
        """
        url = "https://www.bud.hu/budapest_airport/media/hirek"
        headers = {"User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36"}

        try:
            response = requests.get(url, headers=headers, timeout=10)
            if response.status_code == 200:
                soup = BeautifulSoup(response.text, 'html.parser')
                # A repülőtér oldalán a hírek címei általában h2 vagy h3 elemekben vannak
                found = soup.select('h2, .news-list-item__title')
                news = [item.get_text(strip=True) for item in found if len(item.get_text()) > 10]
                if news:
                    return news[:5]
        except Exception as e:
            logger.warning(f"Reptér scraping hiba: {e}")

        # TARTALÉK REPTÉRI INFÓK (Ha a weboldal blokkolna)
        return [
            "✈️ Az összes terminál zavartalanul üzemel.",
            "🛡️ Kérjük, érkezzen 2 órával az indulás előtt.",
            "🚗 A Terminal Parking területén szabad helyek elérhetők.",
            "🛂 Gyorsított biztonsági ellenőrzés (Fast Track) üzemel.",
            "☁️ Időjárás: Repülésre alkalmas, tiszta idő."
        ]