requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.124.4
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.13.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.5.0
pluggy==1.6.0
protobuf==6.33.2
pyarrow==22.0.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
schedule==1.2.2
six==1.17.0
smmap==5.0.2
soupsieve==2.8
SQLAlchemy==2.0.45
starlette==0.50.0
streamlit==1.52.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.38.0
watchdog==6.0.0
